Simplify: dedupe debug-log helpers, drop dead Hyperbee branches
Tier-3 #8 (partial) and Tier-2 #6 (partial) from docs/SIMPLIFICATION_REVIEW.md.

- Extract resolveDebugLogPath + appendDebugLine into src/debug-log.js. They were
  byte-duplicated in orchestrator.js, runtime.js, and storage.js (storage uses
  appendDebugLine at 71 call sites); all three now import the single copy.
- public-channel-bee.js: this.bee is initialized to null and never assigned
  (only this.db is, in _open), so the this.bee?.… branches in getMetadata/
  setMetadata/listVideos were permanent no-ops. Replace them with the equivalent
  empty result, keeping the not-ready guard and the this.bee=null line that a
  test still asserts on.

Verified against the agent's claims: did NOT merge getSwarmStatus shaping
(mobile-handlers vs runtime derive connected/peerCount differently — not
equivalent), and kept applyVideoChanges (used by public-channel-hyperdb test).
